adds a *temp fix* for fractional powers , check later

+++ EggFFI.jl/src/converter.jl
@@ -160,7 +160,8 @@ end
 function build_expr(tree, vars::Dict{String, Num})::Num
     if tree isa String
         n = _leaf_to_number(tree)
-        n !== nothing && return Num(n)
+        # please check here for fractional powers
+        n !== nothing && return Num(Float64(n))
         haskey(NAMED_CONSTANTS, tree) && return Num(NAMED_CONSTANTS[tree])
         return vars[tree]
     end
